backend/requirements.txt
Django==4.2.16
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.0
django-cors-headers==4.3.1
lunarcalendar==0.0.9
uwsgi==2.0.26
psycopg2-binary==2.9.9

